[Fixes #13644] Multilang: update API query using index - improve filtering
