+++ geonode/indexing/management/commands/reindex.py
@@ -99,10 +99,9 @@ class Command(BaseCommand):
                 try:
                     good = None
                     try:
-                        jsonschema = metadata_manager.get_schema()
                         jsoninstance = metadata_manager.build_schema_instance(resource)
                         if not dry_run:
-                            index_manager.update_index(resource.id, jsonschema, jsoninstance)
+                            index_manager.update_index(resource.id, jsoninstance)
                         good = True
 
                     except Exception as e:
